consolidate test folder
