--- hydra/html-templates/layout.tpl
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>{{block "title" .}}{{end}}</title>
-    </head>
-    <body>
-        {{block "authboss" .}}{{end}}
-    </body>
-</html>
